# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la definición de la API del servidor del Hydro Monitor",
        "title": "Hydro Monitor Web API",
        "contact": {
            "name": "Manuel Porto",
            "email": "manu.porto94@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
        "/nodes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Obtiene todos los nodos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_models.NodeDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Crea un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información del nodo",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_models.NodeDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api_models.NodeDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/nodes/{node_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Obtiene la información completa de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.NodeDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza la información de un nodo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Actualiza la información de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Información del nodo",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_models.NodeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "nodes"
                ],
                "summary": "Borra un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/nodes/{node_id}/configuration": {
            "get": {
                "description": "Devuelve un mapa de estados en donde la clave de cada uno es el nombre del mismo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Obtiene la configuración de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/api_models.StateDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Devuelve un mapa de estados (no un array como se ve a continuación) en donde la clave de cada uno es el nombre del mismo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Crea o actualiza la configuración para el nodo dado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuración",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/api_models.StateDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/api_models.StateDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/nodes/{node_id}/manual-reading": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Obtiene el estado de medición manual de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.ManualReadingDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nodes"
                ],
                "summary": "Actualiza el estado de medición manual de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medición manual",
                        "name": "manual_reading",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_models.ManualReadingDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.ManualReadingDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_models.ManualReadingDTO": {
            "type": "object",
            "properties": {
                "manualReading": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "api_models.NodeDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Nodo instalado debajo de un puente"
                },
                "id": {
                    "type": "string",
                    "example": "lujan-1"
                },
                "manualReading": {
                    "type": "boolean",
                    "example": false
                },
                "password": {
                    "type": "string",
                    "example": "aDF23DDcaF45k7J0"
                }
            }
        },
...
        "api_models.StateDTO": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "integer",
                    "example": 25
                },
                "lowerLimit": {
                    "type": "number",
                    "example": 0.5
                },
                "next": {
                    "type": "string",
                    "example": "Alto"
                },
                "picturesNum": {
                    "type": "integer",
                    "example": 1
                },
                "prev": {
                    "type": "string",
                    "example": "Bajo"
                },
                "upperLimit": {
                    "type": "number",
                    "example": 1
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "internal": {
                    "description": "Stores the error returned by an external dependency",
                    "type": "error"
                },
                "message": {
                    "type": "object"
                }
            }
        }
    }
}